compare meshes, unify signatures, and improve performance (#550)
* WIP: start unifying some interfaces and improve performance of structured mesh solver for Euler 2D

* add elixirs used for performance comparisons in 2D

* rename dg/dg -> dg/dg_tree

* add NEWS entries for new mesh types

* unify 2D calling signatures

* unify 1D, 2D, 3D calling signatures

* add 3D mesh comparison benchmarks

* Euler flux in 3D for normal::AbstractVector

* rename new benchmark elixirs consistently

* use get_contravariant_vector to abstract memory layout away

* swap index <-> dimension in the memory layout of contravariant_vectors

* use flux_lax_friedrichs instead of FluxRotated(flux_lax_friedrichs) in Euler elixirs

* specify orientation/normal in many methods

* Apply suggestions from code review


* Apply suggestions from code review


* add some docs

* dispatch only on the mesh (dimension)

* download the mesh file

+++ src/equations/ideal_glm_mhd_1d.jl
@@ -325,7 +325,7 @@ Entropy conserving two-point flux by
   divergence diminishing ideal magnetohydrodynamics equations
   [DOI: 10.1016/j.jcp.2018.03.002](https://doi.org/10.1016/j.jcp.2018.03.002)
 """
-function flux_derigs_etal(u_ll, u_rr, orientation, equations::IdealGlmMhdEquations1D)
+function flux_derigs_etal(u_ll, u_rr, orientation::Integer, equations::IdealGlmMhdEquations1D)
   # Unpack left and right states to get velocities, pressure, and inverse temperature (called beta)
   rho_ll, rho_v1_ll, rho_v2_ll, rho_v3_ll, rho_e_ll, B1_ll, B2_ll, B3_ll = u_ll
   rho_rr, rho_v1_rr, rho_v2_rr, rho_v3_rr, rho_e_rr, B1_rr, B2_rr, B3_rr = u_rr
@@ -383,7 +383,7 @@ end
 
 
 # Calculate maximum wave speed for local Lax-Friedrichs-type dissipation
-@inline function max_abs_speed_naive(u_ll, u_rr, orientation, equations::IdealGlmMhdEquations1D)
+@inline function max_abs_speed_naive(u_ll, u_rr, orientation::Integer, equations::IdealGlmMhdEquations1D)
   rho_ll, rho_v1_ll, rho_v2_ll, rho_v3_ll, rho_e_ll, B1_ll, B2_ll, B3_ll = u_ll
   rho_rr, rho_v1_rr, rho_v2_rr, rho_v3_rr, rho_e_rr, B1_rr, B2_rr, B3_rr = u_rr
 
@@ -413,7 +413,7 @@ Calculate minimum and maximum wave speeds for HLL-type fluxes as in
   An HLLC Riemann solver for magneto-hydrodynamics
   [DOI: 10.1016/j.jcp.2004.08.020](https://doi.org/10.1016/j.jcp.2004.08.020)
 """
-@inline function min_max_speed_naive(u_ll, u_rr, orientation, equations::IdealGlmMhdEquations1D)
+@inline function min_max_speed_naive(u_ll, u_rr, orientation::Integer, equations::IdealGlmMhdEquations1D)
   rho_ll, rho_v1_ll, rho_v2_ll, rho_v3_ll, rho_e_ll, B1_ll, B2_ll, B3_ll = u_ll
   rho_rr, rho_v1_rr, rho_v2_rr, rho_v3_rr, rho_e_rr, B1_rr, B2_rr, B3_rr = u_rr
